+++ backend/src/services/products/GetProductByIdService.ts
@@ -1,13 +1,13 @@
-import Product from "../models/productModel";
+import Product from "../../models/productModel";
 
-class DeleteProductService {
+class GetProductByIdService {
    public async execute(productId: string) {
       const product = await Product.findByPk(productId);
       if (!product) {
          throw new Error("Product not found");
       }
-      await product.destroy();
+      return product;
    }
 }
 
-export default DeleteProductService;
+export default GetProductByIdService;
